modofy delete event process and restore in observer

+++ app/Observers/SerieObserver.php
@@ -10,19 +10,38 @@ class SerieObserver
     public function deleting(Serie $serie)
     {
         // Eliminar registros de las tablas intermedias
-        $serie->platforms()->detach();
-        $serie->actors()->detach();
-        $serie->directors()->detach();
-        $serie->languages()->detach();
+        if ($serie->platforms()->exists()) {
+            $serie->platforms()->detach();
+        }
+        if ($serie->actors()->exists()) {
+            $serie->actors()->detach();
+        }
+        if ($serie->directors()->exists()) {
+            $serie->directors()->detach();
+        }
+        if ($serie->languages()->exists()) {
+            $serie->languages()->detach();
+        }
     }
 
     // Serie restore logic
     public function restoring(Serie $serie)
     {
         // Restaurar registros de las tablas intermedias
-        $serie->platforms()->withTrashed()->restore();
-        $serie->actors()->withTrashed()->restore();
-        $serie->directors()->withTrashed()->restore();
-        $serie->languages()->withTrashed()->restore();
+        $serie->platforms()->withTrashed()->get()->each(function ($platform) {
+            $platform->pivot->restore();
+        });
+
+        $serie->actors()->withTrashed()->get()->each(function ($actor) {
+            $actor->pivot->restore();
+        });
+
+        $serie->directors()->withTrashed()->get()->each(function ($director) {
+            $director->pivot->restore();
+        });
+
+        $serie->languages()->withTrashed()->get()->each(function ($language) {
+            $language->pivot->restore();
+        });
     }
 }
